requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-jazzmin==2.5.0
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
